move macros integration tests
- similar to the runner tests before, these behave like integration
  tests and only use public api
- because of this, move them to their own module
