+++ tests/output/main.rs
@@ -4,4 +4,5 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+mod macros;
 mod runner;
